+++ protected/humhub/docs/CHANGELOG.md
@@ -35,6 +35,7 @@ Important note for LDAP users: There is a new setting "ID Attribute" which shoul
 - Fix: Prevent Integrity check failures.
 - Enh: Added default open content as modal action
 - Enh: Added possibility to add attachments in Notification MailTarget
+- Enh: Added surpressSendToOriginator Notification option
 
 
 1.2.2  (August 2, 2017)